Terraform SDK update, add edit to fields (without delete/recreate) and destination support

+++ kaleido/resource_consortium.go
@@ -16,7 +16,7 @@ package kaleido
 import (
 	"fmt"
 
-	"github.com/hashicorp/terraform/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	kaleido "github.com/kaleido-io/kaleido-sdk-go/kaleido"
 )
 
@@ -24,17 +24,22 @@ func resourceConsortium() *schema.Resource {
 	return &schema.Resource{
 		Create: resourceConsortiumCreate,
 		Read:   resourceConsortiumRead,
+		Update: resourceConsortiumUpdate,
 		Delete: resourceConsortiumDelete,
 		Schema: map[string]*schema.Schema{
 			"name": &schema.Schema{
 				Type:     schema.TypeString,
 				Required: true,
-				ForceNew: true,
 			},
 			"description": &schema.Schema{
 				Type:     schema.TypeString,
 				Required: true,
-				ForceNew: true,
+			},
+			"shared_deployment": &schema.Schema{
+				Type:        schema.TypeBool,
+				Optional:    true,
+				Default:     false,
+				Description: "The decentralized nature of Kaleido means a consortium might be shared with other accounts. When true only create if name does not exist, and delete becomes a no-op.",
 			},
 		},
 	}
@@ -42,18 +47,59 @@ func resourceConsortium() *schema.Resource {
 
 func resourceConsortiumCreate(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(kaleido.KaleidoClient)
-	consortium := kaleido.NewConsortium(d.Get("name").(string),
-		d.Get("description").(string))
+	consortium := kaleido.NewConsortium(
+		d.Get("name").(string),
+		d.Get("description").(string),
+	)
+
+	if d.Get("shared_deployment").(bool) {
+		var consortia []kaleido.Consortium
+		res, err := client.ListConsortium(&consortia)
+		if err != nil {
+			return err
+		}
+		if res.StatusCode() != 200 {
+			return fmt.Errorf("Failed to list existing consortia with status %d: %s", res.StatusCode(), res.String())
+		}
+		for _, c := range consortia {
+			if c.Name == consortium.Name {
+				// Already exists, just re-use
+				d.SetId(c.ID)
+				return resourceConsortiumRead(d, meta)
+			}
+		}
+	}
+
 	res, err := client.CreateConsortium(&consortium)
 	if err != nil {
 		return err
 	}
 	status := res.StatusCode()
 	if status != 201 {
-		return fmt.Errorf("Failed to create consortium with status: %d", status)
+		return fmt.Errorf("Failed to create consortium with status %d", status)
 	}
 
 	d.SetId(consortium.ID)
+
+	return nil
+}
+
+func resourceConsortiumUpdate(d *schema.ResourceData, meta interface{}) error {
+	client := meta.(kaleido.KaleidoClient)
+	consortium := kaleido.NewConsortium(
+		d.Get("name").(string),
+		d.Get("description").(string),
+	)
+	consortiumID := d.Id()
+
+	res, err := client.UpdateConsortium(consortiumID, &consortium)
+	if err != nil {
+		return err
+	}
+	status := res.StatusCode()
+	if status != 200 {
+		return fmt.Errorf("Failed to update consortium %s with status: %d", consortiumID, status)
+	}
 
 	return nil
 }
@@ -75,10 +121,18 @@ func resourceConsortiumRead(d *schema.ResourceData, meta interface{}) error {
 		}
 		return fmt.Errorf("Failed to read consortium with id %s status was: %d, error: %s", d.Id(), status, res.String())
 	}
+	d.Set("name", consortium.Name)
+	d.Set("description", consortium.Description)
 	return nil
 }
 
 func resourceConsortiumDelete(d *schema.ResourceData, meta interface{}) error {
+	if d.Get("shared_deployment").(bool) {
+		// Cannot safely delete if this is shared with other terraform deployments
+		d.SetId("")
+		return nil
+	}
+
 	client := meta.(kaleido.KaleidoClient)
 	res, err := client.DeleteConsortium(d.Id())
